feat: Add new user-level bot setting and functionality
This commit adds a new user-level bot setting to the project. It includes the following changes:
- Added the `config` command to the `setting()` function
- Added a default value to `DEFAULT{}`
- Implemented functionality to handle the new setting in the `maybe_on_message()` function
- Added the new setting to the `get_channel_settings()` function
- Updated the `COMMANDS` list to include the new setting

Additionally, this commit retains empty channel data fields that are not set as default to differentiate when changing defaults. It also stops checking messages if the first message is reached and makes the plugin/fallback flags into channel data and vice versa.